Switch to VS solution with MDK
