+++ Scripts/Usstan's Ice Controller/Instructions.readme
@@ -1,0 +1,4 @@
+﻿---=== Usstan's Ice Controller ===---
+This script was created for use on worlds with water (WaterMod) where pumping ice from the water can quickly result in all cargo containers on the base being full of ice. 
+
+I originally used Coxy's Ice Pump, but wanted a little more control and configuration
