+++ Scripts/Usstan's Ice Controller/MDK/MDK.options.props
@@ -1,0 +1,18 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <!-- 
+    Contains basic settings that should be included in code repositories
+    -->
+    <MDKVersion>1.4.12</MDKVersion>
+    <MDKTrimTypes>
+      <Enabled>yes</Enabled>
+    </MDKTrimTypes>
+    <MDKMinify>
+      <Level>Full</Level>
+    </MDKMinify>
+    <MDKIgnore>
+      <Folder>mdk</Folder>
+    </MDKIgnore>
+  </PropertyGroup>
+</Project>
